// fe/src/lib/api/generated/model/geographicLocationPayload.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { GeographicLocationKindValue } from './geographicLocationKindValue';

export interface GeographicLocationPayload {
  barangay_name_snapshot?: string;
  barangay_psgc_code?: string;
  city_municipality_name_snapshot?: string;
  city_municipality_psgc_code?: string;
  kind: GeographicLocationKindValue;
  not_specified?: boolean;
  province_name_snapshot?: string;
  province_psgc_code?: string;
  region_name_snapshot?: string;
  region_psgc_code?: string;
}